requirements.txt
fastapi==0.115.13
motor==3.7.1
uvicorn==0.34.3
python-jose==3.5.0
passlib==1.7.4
python-dotenv==1.1.0
bcrypt==4.3.0
email-validator==2.2.0